+++ smd/g_trigger.c
@@ -1834,9 +1834,9 @@ int trigger_transition_ents (edict_t *changelevel, edict_t *self)
 			e = G_Spawn();
 			e->classname = gi.TagMalloc(17,TAG_LEVEL);
 			strcpy(e->classname,"info_train_start");
-			e->targetname = gi.TagMalloc(strlen(ent->targetname)+1,TAG_LEVEL);
+			e->targetname = gi.TagMalloc((int)strlen(ent->targetname)+1,TAG_LEVEL);
 			strcpy(e->targetname,ent->targetname);
-			e->target = gi.TagMalloc(strlen(ent->target)+1,TAG_LEVEL);
+			e->target = gi.TagMalloc((int)strlen(ent->target)+1,TAG_LEVEL);
 			strcpy(e->target,ent->target);
 			e->spawnflags = ent->spawnflags;
 			VectorCopy(ent->s.origin,e->s.origin);
